requirements.txt
fastapi
uvicorn[standard]
openai>=2.0.0
python-dotenv
pandas>=2.0.0
openpyxl>=3.1.2
google-genai
google-cloud-storage
google-auth
pdf2image
pillow
